move grpc auth client in its own package test for the grpc server+client using session passes extend jwt library with ED25519 signature implementation remove obsolete interceptor file
